Concatenate XML to PDF report
